+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [0.0.4] - 2026-02-03
+
+### <!-- 0 -->⛰️  Features
+
+- Unify sensor access with a generic Sensor class
+- Add GNSS examples for PTY and TTY output
+
 ## [0.0.2] - 2026-02-02
 
 ### <!-- 0 -->⛰️  Features